code/requirements.txt
cffi==1.15.1
cryptography==41.0.1
ed25519==1.5
psec==1.3.0
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
